// src/viewer/proto/spark_pb.ts
export interface WindowStatistics {
    /**
     * @generated from protobuf field: int32 ticks = 1;
     */
    ticks: number;
    /**
     * @generated from protobuf field: double cpuProcess = 2;
     */
    cpuProcess: number;
    /**
     * @generated from protobuf field: double cpuSystem = 3;
     */
    cpuSystem: number;
    /**
     * @generated from protobuf field: double tps = 4;
     */
    tps: number;
    /**
     * @generated from protobuf field: double msptMedian = 5;
     */
    msptMedian: number;
    /**
     * @generated from protobuf field: double msptMax = 6;
     */
    msptMax: number;
    /**
     * world
     *
     * @generated from protobuf field: int32 players = 7;
     */
    players: number;
    /**
     * @generated from protobuf field: int32 entities = 8;
     */
    entities: number;
    /**
     * @generated from protobuf field: int32 tileEntities = 9;
     */
    tileEntities: number;
    /**
     * @generated from protobuf field: int32 chunks = 10;
     */
    chunks: number;
    /**
     * approximate wall-clock start/end times
     *
     * @generated from protobuf field: int64 startTime = 11;
     */
    startTime: number;
    /**
     * @generated from protobuf field: int64 endTime = 12;
     */
    endTime: number;
    /**
     * @generated from protobuf field: int32 duration = 13;
     */
    duration: number;
}
// ...
export const WindowStatistics = new WindowStatistics$Type();